requirements.txt
click
PyPDF2
python-pptx
openai
python-dotenv
pdf2image
rich